Implement getExpectedNextLeapYear method as a helper method for testing
Had a problem testing getNextLeapYear method since the expected value would change
depending on the year the test was run.
This helper method takes the expected values from a static file,
trying to prevent a possible bug in helper method itself.

I'm aware that there might be better way to test the getNextLeapYear method.
But this is the best solution I've come up with atm.